fix(build): move the random-container helpers out of Helpers.h
Corrects two compile errors in the Phase 1 work, both introduced by
earlier commits on this branch.

1. P1.7 put RandomElement/RandomIndex in Util/Helpers.h and then included
   that from TravelNode.h - a widely included header. Helpers.h also
   declares ltrim/rtrim/trim/split, so every downstream translation unit
   suddenly saw them, and CastCustomSpellAction.cpp defines its own
   `static inline void ltrim(std::string&)`:

     CastCustomSpellAction.cpp:28:20: error: ambiguating new declaration
     of 'void ltrim(std::string&)'

   Fixed at the cause rather than at the collision site: the random
   helpers now live in their own narrow Util/RandomUtils.h and Helpers.h
   is restored untouched. String utilities and container-sampling
   utilities had no business sharing a header, and a narrow header cannot
   leak unrelated declarations into hundreds of TUs.

2. `Capital::bankers` is std::vector<uint16>, not uint32, so the
   RandomElement result in TravelMgr::GetBankerLocations did not bind.
   This is exactly what the old `urand(0, size() - 1)` indexing hid - it
   never had to name the element type.

Also adds the Field.h / QueryResult.h includes QuestBlacklistMgr.cpp
needs; DatabaseEnv.h only forward-declares ResultSet and Field.

Verified: all 11 touched translation units pass -fsyntax-only using their
recorded compile_commands.json flags, and a full `make -j4` links both
worldserver and authserver.

+++ src/Mgr/Travel/TravelMgr.cpp
@@ -10,7 +10,7 @@
 #include "ChatHelper.h"
 #include "Corpse.h"
 #include "Creature.h"
-#include "Helpers.h"
+#include "RandomUtils.h"
 #include "Log.h"
 #include "Map.h"
 #include "MapCollisionData.h"
@@ -4551,7 +4551,7 @@ std::vector<WorldLocation> TravelMgr::GetCityLocations(Player* bot)
     auto const& bankers = selectedCapital->bankers;
     // `bankers` came straight out of the capital definition and was indexed without an emptiness
     // check: a capital with no bankers configured read out of bounds.
-    uint32 const* selectedBankerEntry = RandomElement(bankers);
+    uint16 const* selectedBankerEntry = RandomElement(bankers);
     if (!selectedBankerEntry)
         return fallbackLocations;
     auto locIt = bankerEntryToLocation.find(*selectedBankerEntry);
